Add arch/tls_static
Move arch-dependent static TLS functionality out of kos/thread.
Previously the SH-specific implementation details were hardcoded
into otherwise platform-independent code.
