+++ kernel/arch/dreamcast/kernel/Makefile
@@ -13,7 +13,7 @@ COPYOBJS = banner.o cache.o entry.o irq.o init.o mm.o panic.o
 COPYOBJS += rtc.o timer.o wdt.o perfctr.o perf_monitor.o
 COPYOBJS += init_flags_default.o
 COPYOBJS += mmu.o itlb.o
-COPYOBJS += exec.o execasm.o stack.o gdb_stub.o thdswitch.o arch_exports.o
+COPYOBJS += exec.o execasm.o stack.o gdb_stub.o thdswitch.o tls_static.o arch_exports.o
 COPYOBJS += uname.o
 OBJS = $(COPYOBJS) startup.o
 SUBDIRS =
