+++ kernel/thread/thread.c
@@ -24,9 +24,11 @@
 #include <kos/rwsem.h>
 #include <kos/cond.h>
 #include <kos/genwait.h>
+
 #include <arch/irq.h>
 #include <arch/timer.h>
 #include <arch/arch.h>
+#include <arch/tls_static.h>
 
 /*
 
@@ -40,16 +42,6 @@ BSD kernel quite a bit to get some ideas on priorities, and I am
 also using their queue library verbatim (sys/queue.h).
 
 */
-
-/* TLS Section ELF data - exported from linker script. */
-extern int _tdata_start, _tdata_size;
-extern int _tbss_size;
-extern long _tdata_align, _tbss_align;
-
-/* Utility function for aligning an address or offset. */
-static inline size_t align_to(size_t address, size_t alignment) {
-    return (address + (alignment - 1)) & ~(alignment - 1);
-}
 
 /* Builtin background thread data */
 static alignas(8) uint8_t thd_reaper_stack[512];
@@ -367,94 +359,6 @@ int thd_remove_from_runnable(kthread_t *thd) {
     return 0;
 }
 
-/* Creates and initializes the static TLS segment for a thread,
-   composed of a Thread Control Block (TCB), followed by .TDATA,
-   followed by .TBSS, very carefully ensuring alignment of each
-   subchunk. 
-*/
-static bool thd_create_tls_data(kthread_t *nt) {
-    size_t align, tdata_offset, tdata_end, tbss_offset, 
-        tbss_end, align_rem, tls_size;
-
-    tcbhead_t *tcbhead;
-    void *tdata_segment, *tbss_segment;
-
-    /* Cached and typed local copies of TLS segment data for sizes, 
-       alignments, and initial value data pointer, exported by the 
-       linker script. 
-
-       SIZES MUST BE VOLATILE or the optimizer on non-debug builds will 
-       optimize zero-check conditionals away, since why would the 
-       address of a variable be NULL? (Linker script magic, it can be.)
-   */
-    const volatile size_t   tdata_size  = (size_t)(&_tdata_size);
-    const volatile size_t   tbss_size   = (size_t)(&_tbss_size);
-    const          size_t   tdata_align = tdata_size ? (size_t)_tdata_align : 1;
-    const          size_t   tbss_align  = tbss_size ? (size_t)_tbss_align : 1;
-    const          uint8_t *tdata_start = (const uint8_t *)(&_tdata_start);
-
-    /* Each subsegment of the requested memory chunk must be aligned
-       by the largest segment's memory alignment requirements. 
-   */
-    align = 8;               /* tcbhead_t has to be aligned by 8. */
-    if(tdata_align > align)
-        align = tdata_align; /* .TDATA segment's alignment */
-    if(tbss_align > align)
-        align = tbss_align;  /* .TBSS segment's alignment */
-
-    /* Calculate the sizing and offset location of each subsegment. */
-    tdata_offset = align_to(sizeof(tcbhead_t), align);
-    tdata_end    = tdata_offset + tdata_size;
-    tbss_offset  = align_to(tdata_end, tbss_align);
-    tbss_end     = tbss_offset + tbss_size; 
-
-    /* Calculate final aligned size requirement. */
-    align_rem = tbss_end % align;
-    tls_size  = tbss_end;
-
-    if(align_rem)
-        tls_size += (align - align_rem);
-
-    /* Allocate combined chunk with calculated size and alignment. */
-    tcbhead = aligned_alloc(align, tls_size);
-
-    if(!tcbhead)
-        return false;
-
-    assert(!((uintptr_t)tcbhead % 8)); 
-
-    /* Since we aren't using either member within it, zero out tcbhead. */
-    memset(tcbhead, 0, sizeof(tcbhead_t));
-
-    /* Initialize .TDATA */
-    if(tdata_size) { 
-        tdata_segment = (uint8_t *)tcbhead + tdata_offset;
-
-        /* Verify proper alignment. */
-        assert(!((uintptr_t)tdata_segment % tdata_align));
-
-        /* Initialize tdata_segment with .tdata bytes from ELF. */
-        memcpy(tdata_segment, tdata_start, tdata_size);
-    }
-
-    /* Initialize .TBSS */
-    if(tbss_size) { 
-        tbss_segment = (uint8_t *)tcbhead + tbss_offset; 
-
-        /* Verify proper alignment. */
-        assert(!((uintptr_t)tbss_segment % tbss_align));
-
-        /* Zero-initialize tbss_segment. */
-        memset(tbss_segment, 0, tbss_size);
-    }
-
-    /* Set Thread Pointer and store starting value. */
-    nt->context.gbr = tcbhead;
-    nt->tcbhead = tcbhead;
-
-    return true;
-}
-
 /* New thread function; given a routine address, it will create a
    new kernel thread with the given attributes. When the routine
    returns, the thread will exit. Returns the new thread struct. */
@@ -525,7 +429,7 @@ kthread_t *thd_create_ex(const kthread_attr_t *restrict attr,
                                (uint32_t)thd_birth, params, 0);
 
             /* Create static TLS data */
-            if(!thd_create_tls_data(nt)) {
+            if(!arch_tls_setup_data(nt)) {
                 if(nt->flags & THD_OWNS_STACK)
                     free(nt->stack);
                 free(nt);
@@ -621,7 +525,7 @@ int thd_destroy(kthread_t *thd) {
         free(thd->stack);
 
     /* Free static TLS segment */
-    free(thd->tcbhead);
+    arch_tls_destroy_data(thd);
 
     /* Free the thread */
     free(thd);
@@ -1121,10 +1025,12 @@ int thd_init(void) {
 
     /* Setup a kernel task for the currently running "main" thread */
     kern = thd_create_ex(&kern_attr, NULL, NULL);
-    kern->state = STATE_RUNNING;
+    if(!kern) {
+        dbglog(DBG_DEAD, "thd: failed to create kernel thread\n");
+        return -1;
+    }
 
-    /* Initialize GBR register for Main Thread */
-    __builtin_set_thread_pointer((void*)kern->context.gbr);
+    kern->state = STATE_RUNNING;
 
     /* De-scehdule the thread (it's STATE_RUNNING) */
     thd_remove_from_runnable(kern);
@@ -1141,6 +1047,9 @@ int thd_init(void) {
     /* Main thread -- the kern thread */
     thd_current = kern;
     irq_set_context(&kern->context);
+
+    /* Initialize tls */
+    arch_tls_init();
 
     thd_update_cpu_time(thd_current);
